ELMU #147 Simplify and rename user domain module
